reorganized nav bar, moved friends information on user page to a new friend list page
