+++ react-app/src/components/NavBar.js
@@ -1,9 +1,11 @@
-
 import React from 'react';
+import { useSelector } from 'react-redux'
 import { NavLink } from 'react-router-dom';
 import LogoutButton from './auth/LogoutButton';
 
 const NavBar = () => {
+  const user = useSelector(state => state.session.user)
+
   return (
     <nav>
       <ul>
@@ -12,16 +14,22 @@ const NavBar = () => {
             Home
           </NavLink>
         </li>
-        <li>
-          <NavLink to='/login' exact={true} activeClassName='active'>
-            Login
-          </NavLink>
-        </li>
-        <li>
-          <NavLink to='/sign-up' exact={true} activeClassName='active'>
-            Sign Up
-          </NavLink>
-        </li>
+        {!user && (
+          <li>
+            <NavLink to='/login' exact={true} activeClassName='active'>
+              Login
+            </NavLink>
+          </li>
+        )}
+        {!user && (
+          <li>
+            <NavLink to='/sign-up' exact={true} activeClassName='active'>
+              Sign Up
+            </NavLink>
+          </li>
+        )}
+        {user && (
+        <>
         <li>
           <NavLink to='/users' exact={true} activeClassName='active'>
             Users
@@ -33,6 +41,11 @@ const NavBar = () => {
           </NavLink>
         </li>
         <li>
+          <NavLink to='/my-friends' exact={true} activeClassName='active'>
+            Friends List
+          </NavLink>
+        </li>
+        <li>
           <NavLink to='/new-challenge' exact={true} activeClassName='active'>
             Send Challenge to a Friend
           </NavLink>
@@ -40,6 +53,8 @@ const NavBar = () => {
         <li>
           <LogoutButton />
         </li>
+        </>
+        )}
       </ul>
     </nav>
   );
